Refactor Saga class to use asynchronous methods
Replaced `CheckForCompletion` with `CheckForCompletionAsync` in the `Saga` class, updating its definition to return a `Task`. Updated calls to `TelemetryAdapter.RecordCustomEventAsync` and `TelemetryAdapter.EndSagaAsync` to use `await`. Added a `finally` block to ensure all possible reminders are canceled if the saga is done. Modified `ReportReminderAsync` to call `CancelReminderAsync` if the saga state is corrupted. Updated calls to `CheckForCompletion` in `SagaAction.cs` and `SagaRevertAction.cs` to `await _saga.CheckForCompletionAsync` and `await Saga.CheckForCompletionAsync`, respectively.