Update README with Saga Log state store management
Added a new section in README.md titled "Sagaway Saga Log state store" to guide on managing saga logs for long-running sagas. The update includes methods to reduce overhead on the Dapr Actor state store by ignoring the SagaLog or using an external state store. Introduced the `IStepRecorder` interface for custom implementations and explained the use of `WithNullStepRecorder` and `WithStepRecorder` in the Saga Builder. Provided a code example for registering `DaprStoreStepRecorder` in the DI container and using it with `WithStepRecorder`.